backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
motor==3.5.3
pymongo==4.7.3
python-dotenv==1.0.1
pydantic==2.10.4
